virt-launcher, hostdev: Extract hot(un)plug from SR-IOV
Extract the hot(un)plug code from the SR-IOV package to allow its usage
by the other hostdevice types.
