+++ pkg/virt-launcher/virtwrap/device/hostdevice/BUILD.bazel
@@ -5,6 +5,7 @@ go_library(
     srcs = [
         "addresspool.go",
         "hostdev.go",
+        "hotplug.go",
     ],
     importpath = "kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/device/hostdevice",
     visibility = ["//visibility:public"],
@@ -14,6 +15,7 @@ go_library(
         "//pkg/virt-launcher/virtwrap/device:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
+        "//vendor/libvirt.org/go/libvirt:go_default_library",
     ],
 )
 
@@ -23,6 +25,7 @@ go_test(
         "addresspool_test.go",
         "hostdev_test.go",
         "hostdevice_suite_test.go",
+        "hotplug_test.go",
     ],
     deps = [
         ":go_default_library",
@@ -32,5 +35,6 @@ go_test(
         "//vendor/github.com/onsi/ginkgo:go_default_library",
         "//vendor/github.com/onsi/ginkgo/extensions/table:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
+        "//vendor/libvirt.org/go/libvirt:go_default_library",
     ],
 )
