+++ pkg/virt-launcher/virtwrap/device/hostdevice/sriov/hostdev.go
@@ -20,15 +20,10 @@
 package sriov
 
 import (
-	"encoding/xml"
 	"fmt"
-	"strings"
 	"time"
 
-	"libvirt.org/go/libvirt"
-
 	v1 "kubevirt.io/api/core/v1"
-	"kubevirt.io/client-go/log"
 	"kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/api"
 	"kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/device"
 	"kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/device/hostdevice"
@@ -36,10 +31,6 @@ import (
 
 const (
 	AliasPrefix = "sriov-"
-
-	MaxConcurrentHotPlugDevicesEvents = 32
-
-	affectLiveAndConfigLibvirtFlags = libvirt.DOMAIN_DEVICE_MODIFY_LIVE | libvirt.DOMAIN_DEVICE_MODIFY_CONFIG
 )
 
 func CreateHostDevices(vmi *v1.VirtualMachineInstance) ([]api.HostDevice, error) {
@@ -78,144 +69,9 @@ func createHostDevicesMetadata(ifaces []v1.Interface) []hostdevice.HostDeviceMet
 	return hostDevicesMetaData
 }
 
-type deviceDetacher interface {
-	DetachDeviceFlags(xml string, flags libvirt.DomainDeviceModifyFlags) error
-}
-
-type eventRegistrar interface {
-	Register() error
-	Deregister() error
-	EventChannel() <-chan interface{}
-}
-
-func SafelyDetachHostDevices(domainSpec *api.DomainSpec, eventDetach eventRegistrar, dom deviceDetacher, timeout time.Duration) error {
-	sriovDevices := FilterHostDevices(domainSpec)
-	if len(sriovDevices) == 0 {
-		log.Log.Info("No SR-IOV host-devices to detach.")
-		return nil
-	}
-
-	if err := eventDetach.Register(); err != nil {
-		return fmt.Errorf("failed to detach host-devices: %v", err)
-	}
-	defer func() {
-		if err := eventDetach.Deregister(); err != nil {
-			log.Log.Reason(err).Errorf("failed to detach host-devices: %v", err)
-		}
-	}()
-
-	if err := detachHostDevices(dom, sriovDevices); err != nil {
-		return err
-	}
-
-	return waitHostDevicesToDetach(eventDetach, sriovDevices, timeout)
-}
-
-func FilterHostDevices(domainSpec *api.DomainSpec) []api.HostDevice {
-	var hostDevices []api.HostDevice
-
-	for _, hostDevice := range domainSpec.Devices.HostDevices {
-		if hostDevice.Alias != nil && strings.HasPrefix(hostDevice.Alias.GetName(), AliasPrefix) {
-			hostDevices = append(hostDevices, hostDevice)
-		}
-	}
-	return hostDevices
-}
-
-func detachHostDevices(dom deviceDetacher, hostDevices []api.HostDevice) error {
-	for _, hostDev := range hostDevices {
-		devXML, err := xml.Marshal(hostDev)
-		if err != nil {
-			return fmt.Errorf("failed to encode (xml) hostdev %v, err: %v", hostDev, err)
-		}
-		err = dom.DetachDeviceFlags(string(devXML), affectLiveAndConfigLibvirtFlags)
-		if err != nil {
-			return fmt.Errorf("failed to detach hostdev %s, err: %v", devXML, err)
-		}
-		log.Log.Infof("Successfully hot-unplug hostdev: %s (%v)", hostDev.Alias.GetName(), hostDev.Source.Address)
-	}
-	return nil
-}
-
-func waitHostDevicesToDetach(eventDetach eventRegistrar, hostDevices []api.HostDevice, timeout time.Duration) error {
-	var detachedHostDevices []string
-	var desiredDetachCount = len(hostDevices)
-
-	for {
-		select {
-		case deviceAlias := <-eventDetach.EventChannel():
-			if dev := deviceLookup(hostDevices, deviceAlias.(string)); dev != nil {
-				detachedHostDevices = append(detachedHostDevices, dev.Alias.GetName())
-			}
-			if desiredDetachCount == len(detachedHostDevices) {
-				return nil
-			}
-		case <-time.After(timeout):
-
-			return fmt.Errorf(
-				"failed to wait for host-devices detach, timeout reached: %v/%v",
-				detachedHostDevices, hostDevicesNames(hostDevices))
-		}
-	}
-}
-
-func hostDevicesNames(hostDevices []api.HostDevice) []string {
-	var names []string
-	for _, dev := range hostDevices {
-		names = append(names, dev.Alias.GetName())
-	}
-	return names
-}
-
-func deviceLookup(hostDevices []api.HostDevice, deviceAlias string) *api.HostDevice {
-	deviceAlias = strings.TrimPrefix(deviceAlias, api.UserAliasPrefix)
-	for _, dev := range hostDevices {
-		if dev.Alias.GetName() == deviceAlias {
-			return &dev
-		}
-	}
-	return nil
-}
-
-type deviceAttacher interface {
-	AttachDeviceFlags(xmlData string, flags libvirt.DomainDeviceModifyFlags) error
-}
-
-func AttachHostDevices(dom deviceAttacher, hostDevices []api.HostDevice) error {
-	var errs []error
-	for _, hostDev := range hostDevices {
-		if err := attachHostDevice(dom, hostDev); err != nil {
-			errs = append(errs, err)
-		}
-	}
-
-	if len(errs) > 0 {
-		return buildAttachHostDevicesErrorMessage(errs)
-	}
-
-	return nil
-}
-
-func attachHostDevice(dom deviceAttacher, hostDev api.HostDevice) error {
-	devXML, err := xml.Marshal(hostDev)
-	if err != nil {
-		return fmt.Errorf("failed to encode (xml) host-device %v, err: %v", hostDev, err)
-	}
-	err = dom.AttachDeviceFlags(string(devXML), affectLiveAndConfigLibvirtFlags)
-	if err != nil {
-		return fmt.Errorf("failed to attach host-device %s, err: %v", devXML, err)
-	}
-	log.Log.Infof("Successfully hot-plug host-device: %s (%v)", hostDev.Alias.GetName(), hostDev.Source.Address)
-
-	return nil
-}
-
-func buildAttachHostDevicesErrorMessage(errors []error) error {
-	errorMessageBuilder := strings.Builder{}
-	for _, err := range errors {
-		errorMessageBuilder.WriteString(err.Error() + "\n")
-	}
-	return fmt.Errorf(errorMessageBuilder.String())
+func SafelyDetachHostDevices(domainSpec *api.DomainSpec, eventDetach hostdevice.EventRegistrar, dom hostdevice.DeviceDetacher, timeout time.Duration) error {
+	sriovDevices := hostdevice.FilterHostDevicesByAlias(domainSpec, AliasPrefix)
+	return hostdevice.SafelyDetachHostDevices(sriovDevices, eventDetach, dom, timeout)
 }
 
 func GetHostDevicesToAttach(vmi *v1.VirtualMachineInstance, domainSpec *api.DomainSpec) ([]api.HostDevice, error) {
@@ -223,27 +79,9 @@ func GetHostDevicesToAttach(vmi *v1.VirtualMachineInstance, domainSpec *api.Doma
 	if err != nil {
 		return nil, err
 	}
-	currentAttachedSRIOVHostDevices := FilterHostDevices(domainSpec)
+	currentAttachedSRIOVHostDevices := hostdevice.FilterHostDevicesByAlias(domainSpec, AliasPrefix)
 
-	sriovHostDevicesToAttach := DifferenceHostDevicesByAlias(sriovDevices, currentAttachedSRIOVHostDevices)
+	sriovHostDevicesToAttach := hostdevice.DifferenceHostDevicesByAlias(sriovDevices, currentAttachedSRIOVHostDevices)
 
 	return sriovHostDevicesToAttach, nil
-}
-
-// DifferenceHostDevicesByAlias given two slices of host-devices, according to Alias.Name,
-// it returns a slice with host-devices that exists on the first slice and not exists on the second.
-func DifferenceHostDevicesByAlias(desiredHostDevices, actualHostDevices []api.HostDevice) []api.HostDevice {
-	actualHostDevicesByAlias := make(map[string]struct{}, len(actualHostDevices))
-	for _, hostDev := range actualHostDevices {
-		actualHostDevicesByAlias[hostDev.Alias.GetName()] = struct{}{}
-	}
-
-	var filteredSlice []api.HostDevice
-	for _, desiredHostDevice := range desiredHostDevices {
-		if _, exists := actualHostDevicesByAlias[desiredHostDevice.Alias.GetName()]; !exists {
-			filteredSlice = append(filteredSlice, desiredHostDevice)
-		}
-	}
-
-	return filteredSlice
 }
